+++ src/main/java/com/ppp/billing/model/JobColorCombination.java
@@ -29,7 +29,7 @@ public class JobColorCombination {
 	private int backColorNumber; 
 	
 	@Column(name="number_of_signature")
-	private int numberOfSignature; 
+	private double numberOfSignature; 
 	
 	@ManyToOne
 	@JoinColumn(name = "print_type_id", referencedColumnName = "id")
